requirements.txt
fastapi
uvicorn
rasterio
numpy
pandas
scikit-learn
geopandas
rio-tiler
pillow